# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Freelance Fix API
  description: >
    Middle Tier API powering the localized gig-based platform.
    Integrates with a PostgreSQL database via Prisma.
  version: 2.0.0
servers:
  - url: http://localhost:3001/api
    description: Local Express Backend

# ---------------------------------------------------------------------------
# Security
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Reusable Schemas (derived from prisma/schema.prisma)
  # -------------------------------------------------------------------------
  schemas:
...
    User:
      type: object
      properties:
        id:         { type: string, format: uuid }
        email:      { type: string, format: email }
        first_name: { type: string, nullable: true }
        last_name:  { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        freelancer: { $ref: '#/components/schemas/Freelancer', nullable: true }
        client:     { $ref: '#/components/schemas/Client',     nullable: true }
    Freelancer:
      type: object
      properties:
        id:            { type: string, format: uuid }
        bio:           { type: string, nullable: true }
        portfolio_url: { type: string, nullable: true }
        hourly_rate:   { type: number, nullable: true, example: 75.00 }
        created_at:    { type: string, format: date-time }
        user:
          type: object
          properties:
            id:         { type: string, format: uuid }
            first_name: { type: string, nullable: true }
            last_name:  { type: string, nullable: true }
            email:      { type: string, format: email }
    Client:
      type: object
      properties:
        id:               { type: string, format: uuid }
        company_name:     { type: string, nullable: true }
        spending_history: { type: number, example: 0.00 }
        created_at:       { type: string, format: date-time }
...
    GigPackage:
      type: object
      properties:
        id:            { type: integer }
        gig_id:        { type: integer }
        tier:          { type: string, enum: [Basic, Standard, Premium] }
        price:         { type: number, example: 99.00 }
        description:   { type: string }
        delivery_days: { type: integer, example: 5 }
        created_at:    { type: string, format: date-time }
    Order:
      type: object
      properties:
        id:             { type: integer }
        client_id:      { type: string, format: uuid }
        gig_package_id: { type: integer }
        status:         { type: string, enum: [Pending, In_Progress, Completed, Cancelled] }
        total_amount:   { type: number }
        due_date:       { type: string, format: date-time, nullable: true }
        created_at:     { type: string, format: date-time }
        gig_package:    { $ref: '#/components/schemas/GigPackage', nullable: true }
        payment:        { $ref: '#/components/schemas/Payment',    nullable: true }
        review:         { $ref: '#/components/schemas/Review',     nullable: true }
    Payment:
      type: object
      properties:
        id:             { type: integer }
        order_id:       { type: integer }
        amount:         { type: number }
        status:         { type: string, enum: [Pending, Escrow, Released, Refunded] }
        payment_method: { type: string, nullable: true }
        processed_at:   { type: string, format: date-time, nullable: true }
        created_at:     { type: string, format: date-time }
    Review:
      type: object
      properties:
        id:         { type: integer }
        order_id:   { type: integer }
        rating:     { type: integer, minimum: 1, maximum: 5 }
        comment:    { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    Message:
      type: object
      properties:
        id:          { type: integer }
        sender_id:   { type: string, format: uuid }
        receiver_id: { type: string, format: uuid }
        content:     { type: string }
        is_read:     { type: boolean }
        created_at:  { type: string, format: date-time }
        sender:
          type: object
          properties:
            id:         { type: string, format: uuid }
            first_name: { type: string, nullable: true }
            last_name:  { type: string, nullable: true }
        receiver:
          type: object
          properties:
            id:         { type: string, format: uuid }
            first_name: { type: string, nullable: true }
            last_name:  { type: string, nullable: true }
...
    Error:
      type: object
      properties:
        error: { type: string, example: "Something went wrong" }

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /users/profile:
    get:
      operationId: getProfile
      tags: [Users]
      summary: Get the authenticated user's profile (includes freelancer/client sub-profile)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User object with nested freelancer and/or client profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    put:
      operationId: updateProfile
      tags: [Users]
      summary: Update profile — creates or upserts freelancer/client sub-profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:    { type: string }
                last_name:     { type: string }
                role:          { type: string, enum: [freelancer, client], description: "Explicitly set the role to create" }
                bio:           { type: string, description: "Freelancer only" }
                portfolio_url: { type: string, description: "Freelancer only" }
                hourly_rate:   { type: number, description: "Freelancer only" }
                company_name:  { type: string, description: "Client only" }
      responses:
        '200':
          description: Updated user object with nested profiles
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # --- Freelancers (public search) ------------------------------------------
  /freelancers:
    get:
      operationId: searchFreelancers
      tags: [Freelancers]
      summary: Search freelancer profiles
      security: []
      parameters:
        - in: query
          name: query
          schema: { type: string }
          description: Text search against first_name, last_name, and bio
        - in: query
          name: max_hourly_rate
          schema: { type: number }
          description: Return only freelancers at or below this hourly rate
      responses:
        '200':
          description: Array of freelancer profiles (max 50)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Freelancer' }
        '4XX':
          description: Client error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '5XX':
          description: Server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # --- Gigs -----------------------------------------------------------------
...
  /orders:
    get:
      operationId: getOrders
      tags: [Orders]
      summary: Get all orders for the authenticated user (as client or freelancer)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of orders with nested gig_package, payment, and review
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    post:
      operationId: createOrder
      tags: [Orders]
      summary: Purchase a gig package — creates Order and Escrow Payment (Clients only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gig_package_id]
              properties:
                gig_package_id: { type: integer }
      responses:
        '201':
          description: Order created with linked payment in Escrow status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400':
          description: Missing gig_package_id or package not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: User does not have a client profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /orders/{orderId}/status:
    parameters:
      - name: orderId
        in: path
        required: true
        schema: { type: integer }

    patch:
      operationId: updateOrderStatus
      tags: [Orders]
      summary: Update order status — completing an order releases the Escrow payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [Pending, In_Progress, Completed, Cancelled] }
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400':
          description: Invalid status value
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Caller is not part of this order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Order not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /orders/{orderId}/reviews:
    parameters:
      - name: orderId
        in: path
        required: true
        schema: { type: integer }

    post:
      operationId: postReview
      tags: [Orders]
      summary: Post a review for a completed order (Client who placed the order only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating:  { type: integer, minimum: 1, maximum: 5 }
                comment: { type: string }
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
        '400':
          description: Invalid rating or order is not Completed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Caller is not the client of this order
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: This order already has a review
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # --- Briefs ---------------------------------------------------------------
...
  /messages:
    get:
      operationId: getMessages
      tags: [Messages]
      summary: Get all messages where the user is sender or receiver (newest first, max 200)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of messages with sender and receiver name info
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Message' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    post:
      operationId: sendMessage
      tags: [Messages]
      summary: Send a message to another user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [receiver_id, content]
              properties:
                receiver_id: { type: string, format: uuid }
                content:     { type: string }
      responses:
        '201':
          description: Message sent
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }
        '400':
          description: Missing receiver_id or content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Receiver user not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }